refactor service 패키지에 속했던 클래스들을 domain 패키지 하위로 이동
