+++ src/main/java/racing/domain/RacingGame.java
@@ -1,4 +1,4 @@
-package racing.service;
+package racing.domain;
 
 import racing.domain.Car;
 import racing.domain.Cars;
