+++ src/test/java/racing/domain/CarTest.java
@@ -2,8 +2,6 @@ package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racing.service.GameRule;
-import racing.service.RandomGameRule;
 
 import static org.assertj.core.api.Assertions.assertThat;
